@@ -1,12 +1,48 @@
+/*
+ * Copyright (c) 2019 Tobias Teumert
+ * All rights reserved.
+ * 
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ * 
+ * 1. Redistributions of source code must retain the above copyright notice, this
+ *    list of conditions and the following disclaimer. 
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ * 
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
 package data.model;
-
+/**
+ * Data-Class<p>
+ * Contains two Strings with the short Area-Code and the corresponding Area-Name
+ * @author Tobias
+ *
+ */
 public class Area 
 {
-	private String areaCode,areaName;
+	/**
+	 * Contains the Area-Code
+	 */
+	private String areaCode;
+	/**
+	 * Contains the Area-Name
+	 */
+	private String areaName;
 	/**
 	 * 
-	 * @param s1 Abkürzung
-	 * @param s2 Auschgeschriebener Area Name
+	 * @param s1 Shorthand Area Code as String
+	 * @param s2 Complete Area name as String
 	 */
 	public Area(String s1,String s2)
 	{
@@ -15,14 +51,14 @@ public class Area
 	}
 
 	/**
-	 * @return the shortArea
+	 * @return Getter for short Area Code
 	 */
 	public String getAreaCode() {
 		return areaCode;
 	}
 
 	/**
-	 * @return the longArea
+	 * @return Getter for corresponding Area Name
 	 */
 	public String getAreaName() {
 		return areaName;
