@@ -1,3 +1,27 @@
+/*
+ * Copyright (c) 2019 Tobias Teumert
+ * All rights reserved.
+ * 
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ * 
+ * 1. Redistributions of source code must retain the above copyright notice, this
+ *    list of conditions and the following disclaimer. 
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ * 
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
 package data.model;
 
 import java.time.LocalDateTime;
@@ -11,24 +35,50 @@ import data.model.Reading.LevelType;
  */
 public class ReadingKey 
 {
-	StationId id;
-	LocalDateTime date;
-	LevelType type;
+	/**
+	 * The StationID
+	 */
+	private StationId id;
+	/**
+	 * The Date of the reading
+	 */
+	private LocalDateTime date;
+	/**
+	 * The Leveltype of the Reading
+	 */
+	private LevelType type;
 	
+	/**
+	 * Constructs a Reading Key from given values
+	 * @param id	The StationID
+	 * @param ldt	The Date
+	 * @param type	The Leveltype
+	 */
 	public ReadingKey(StationId id, LocalDateTime ldt, LevelType type) {
 		this.id = id;
 		this.date = ldt;
 		this.type = type;
 	}
 	
+	/**
+	 * Getter for Station ID
+	 * @return the StationID
+	 */
 	public StationId getId() {
 		return id;
 	}
 	
+	/**
+	 * Getter for Leveltype
+	 * @return the Leveltype
+	 */
 	public LevelType getType() {
 		return type;
 	}
-	
+	/**
+	 * Getter for the Date
+	 * @return The Date
+	 */
 	public LocalDateTime getDate() {
 		return date;
 	}
