@@ -1,25 +1,51 @@
+/*
+ * Copyright (c) 2019 Tobias Teumert
+ * All rights reserved.
+ * 
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ * 
+ * 1. Redistributions of source code must retain the above copyright notice, this
+ *    list of conditions and the following disclaimer. 
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ * 
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
 package data.model;
 
 /**
- * Stationsklasse für die Speicherung der Stationsdaten
+ * Representation of a physical station which generates the sounding data by atmospheric probe starts (0 UTC and 12 UTC)<p>
+ * Contains longitude and latitude in decimals rather then degrees
  * @author Tobias
  *
  */
 public class Station {
 
+	/** Unique Station ID */
 	private StationId stationId;
 	/** Name der Station */
 	private String stationName;	
-	/** longitude */
+	/** The longitude of the station in decimals*/
 	private double longitude;
-	/** */
+	/** The latitude of the station in decimals*/
 	private double latitude;	
-	/**Elevation of Station */
+	/** The elevation of station in meters */
 	private int elevation;	
-	/**ICAO Code */
+	/**The ICAO Code */
 	private String icao;
-	/**Country Code */
-	private String country;//REGEX NOCH PUTT geht noch nicht
+	/**The Country Code */
+	private String country;//FIXME REGEX NOCH PUTT geht noch nicht
 	
 	/*public Station()
 	{
@@ -31,7 +57,7 @@ public class Station {
 	 * @param stationName 	Name of Station
 	 * @param longi 		Longitude in Decimals
 	 * @param lati			Latitude in Decimals
-	 * @param elevation		elevation in M
+	 * @param elevation		elevation in meters
 	 * @param icao			International ICAO Code if applicable
 	 */
 	public Station(StationId stationID, String stationName, double longi,double lati, int elevation, String icao)
@@ -49,6 +75,7 @@ public class Station {
 	}
 	
 	/**
+	 * Getter for station name
 	 * @return the stationName
 	 */
 	public String getStationName() {
@@ -56,6 +83,7 @@ public class Station {
 	}
 
 	/**
+	 * Setter for Station Name
 	 * @param stationName the stationName to set
 	 */
 	public void setStationName(String stationName) {
@@ -63,6 +91,7 @@ public class Station {
 	}
 
 	/**
+	 * Getter for Station ID
 	 * @return the stationID
 	 */
 	public StationId getStationID() {
@@ -70,6 +99,7 @@ public class Station {
 	}
 
 	/**
+	 * Setter for Station ID
 	 * @param stationID the stationID to set
 	 */
 	public void setStationID(StationId stationID) {
@@ -77,6 +107,7 @@ public class Station {
 	}
 	
 	/**
+	 * getter for Longitude
 	 * @return the longitude
 	 */
 	public double getLongitude() {
@@ -84,6 +115,7 @@ public class Station {
 	}
 
 	/**
+	 * Setter for longitude
 	 * @param longitude the longitude to set
 	 */
 	public void setLongitude(double longitude) {
@@ -91,6 +123,7 @@ public class Station {
 	}
 
 	/**
+	 * Getter for Latitude
 	 * @return the latitude
 	 */
 	public double getLatitude() {
@@ -98,6 +131,7 @@ public class Station {
 	}
 
 	/**
+	 * Setter for Latitude
 	 * @param latitude the latitude to set
 	 */
 	public void setLatitude(double latitude) {
@@ -105,6 +139,7 @@ public class Station {
 	}
 
 	/**
+	 * Getter for the elevation
 	 * @return the elevation
 	 */
 	public int getElevation() {
@@ -112,6 +147,7 @@ public class Station {
 	}
 
 	/**
+	 * Setter for the elevation of the station
 	 * @param elevation the elevation to set
 	 */
 	public void setElevation(int elevation) {
@@ -119,6 +155,7 @@ public class Station {
 	}
 
 	/**
+	 * Getter for ICAO Code
 	 * @return the stationCode
 	 */
 	public String getIcao() {
@@ -126,12 +163,14 @@ public class Station {
 	}
 
 	/**
+	 * Setter for ICAO Code
 	 * @param stationCode the ICAO Code to set
 	 */
 	public void setIcao(String stationCode) {
 		this.icao = stationCode;
 	}
 	/**
+	 * Getter for Country Code
 	 * @return the country
 	 */
 	public String getCountry() {
@@ -139,6 +178,7 @@ public class Station {
 	}
 
 	/**
+	 * Setter for Country Code
 	 * @param country the country to set
 	 */
 	public void setCountry(String country) {
