@@ -6,36 +6,42 @@ public class WeatherDataParserValidationResult
 	private Sounding data;
 	private Exception exceptions;
 	/**
+	 * Validity of parsed data
 	 * @return the isValid
 	 */
 	public boolean isValid() {
 		return isValid;
 	}
 	/**
+	 * Setter for the validity
 	 * @param isValid the isValid to set
 	 */
 	public void setValid(boolean isValid) {
 		this.isValid = isValid;
 	}
 	/**
-	 * @return the data
+	 * Getter for the Sounding
+	 * @return the parsed data
 	 */
 	public Sounding getData() {
 		return data;
 	}
 	/**
+	 * Setter for Sounding
 	 * @param data the data to set
 	 */
 	public void setData(Sounding data) {
 		this.data = data;
 	}
 	/**
-	 * @return the exceptions
+	 * Getter for associated exception
+	 * @return the exception
 	 */
 	public Exception getException() {
 		return exceptions;
 	}
 	/**
+	 * setter for the exception
 	 * @param exceptions the exceptions to set
 	 */
 	public void setException(Exception exceptions) {
