@@ -1,6 +1,31 @@
+/*
+ * Copyright (c) 2019 Tobias Teumert
+ * All rights reserved.
+ * 
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ * 
+ * 1. Redistributions of source code must retain the above copyright notice, this
+ *    list of conditions and the following disclaimer. 
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ * 
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
 package data.provider;
 
 import java.io.BufferedReader;
+import java.io.BufferedWriter;
 import java.io.FileNotFoundException;
 import java.io.FileReader;
 import java.io.IOException;
@@ -20,13 +45,19 @@ import data.model.StationId;
 import data.model.WeatherDataParser;
 import data.model.WeatherDataParserValidationResult;
 
+/**
+ * The HDD data provider<BR>
+ * Extends ChainedDataProvider to be able to get data by another Provider if not locally stored 
+ * @author Tobias
+ *
+ */
 public class LocalHddDataProvider extends ChainedDataProvider{
 
 	WeatherDataParser parser;
 	
 	/**
 	 * 
-	 * File Structure property\\$region_Code\\$Station_Code\\$LocalDateTime.txt
+	 * File Structure <br><code>property\\$region_Code\\$Station_Code\\$LocalDateTime.txt</code>
 	 * 
 	 * @param provider
 	 * @param prop
@@ -46,38 +77,35 @@ public class LocalHddDataProvider extends ChainedDataProvider{
 	@Override
 	public Sounding getSounding(StationId station, LocalDateTime time) {
 		//Load Sounding with parser from HDD or push it one up to webprovider
-		//System.out.println("Im HDD Provider hier muss was passieren");
+		
 		String datadir = ""+properties.get("data_dir");
-		//System.out.println("DataDir = "+datadir);
 		
-		//System.out.println("Station ="+station);
-		
+		//create Path to look on HDD
 		Path regionP = Path.of(datadir)
 				.resolve(""+station.getArea().getAreaCode())
 				.resolve(""+station.getStationID())
 				.resolve(""+time.toLocalDate().format(DateTimeFormatter.ISO_DATE)+"T"+time.getHour()+".txt");
 		
-		if(Files.exists(regionP))//Region existiert
+		if(Files.exists(regionP))//Region Path exists
 		{
-			System.out.println("File Exists");
+			
 			
 			try (BufferedReader br =
 	                   new BufferedReader(new FileReader(regionP.toFile(),Charset.forName("UTF-8")))){
 				
 				List<String> tokens = br.lines().collect(Collectors.toList());
-				//System.out.println("Tokens aus dem reader in hdd read");
-				//tokens.forEach(System.out::println);
+				
 				WeatherDataParserValidationResult result = parser.validate(tokens.stream().collect(Collectors.joining("\r\n")), station);
 				
 				if (result.isValid()) {
-					System.out.println("Vaild Result from parser");
-					// soundingsByStation.put(station, result.getData());
+					//FIXME Rewrite to Logger
+					System.out.println("HDD - Vaild Result from parser");
+					//soundingsByStation.put(station, result.getData());
 					stations.add(result.getData().getStation());
 					return result.getData();
 				} else {
-					System.out.println("No Valid Result from Parser");
-					System.out.println(result.getException());
-					// result.getException().printStackTrace();
+					System.out.println("HDD - No Valid Result from Parser");
+					System.out.println(result.getException());				
 					result.getData().setDateAndTime(time);
 					return result.getData();
 				}
@@ -88,13 +116,17 @@ public class LocalHddDataProvider extends ChainedDataProvider{
 			} catch (IOException e) {
 				
 				e.printStackTrace();
-			}		
+			}
+			// if local data is corrupted use this to get new from upstream
+			System.out.println("HDD - Upstream");
 			return upstream.getSounding(station, time);
 		}
-		else
+		else // no data on HDD get from upstream
+		{
+			System.out.println("Get from web");
 			return upstream.getSounding(station, time);
-		
-		//return null;
+			
+		}
 	}
 
 	@Override
@@ -102,8 +134,43 @@ public class LocalHddDataProvider extends ChainedDataProvider{
 			String icao) {
 		
 		Station station = new Station(stationID, stationName, longi, lati, elevation, icao);
+		//checks wether the station is already contained in the stations-Set
 		return stations.stream().filter(arg -> arg.equals(station)).findFirst().orElse(station);
 		
 	}
+	/**
+	 * Writes a string containing a sounding to a file on HDD
+	 * @param raw	Raw string to be saved 
+	 * @param ldt	time to construct path
+	 * @param id	id of station to construct path
+	 */
+	public void cacheRawSounding(String raw,LocalDateTime ldt,StationId id)
+	{
+		String datadir = ""+properties.get("data_dir");
+		//Construct path to save to HDD
+		Path toSave = Path.of(datadir)
+				.resolve(""+id.getArea().getAreaCode())
+				.resolve(""+id.getStationID())
+				.resolve(""+ldt.toLocalDate().format(DateTimeFormatter.ISO_DATE)+"T"+ldt.getHour()+".txt");
+		//FIXME Try-close block
+		try {
+			
+			Path p1 = Path.of(datadir).resolve(id.getArea().getAreaCode());
+			if(!Files.exists(p1)) //Checks if Directory of area exists
+				Files.createDirectories(p1);
+			Path p2 = p1.resolve(""+id.getStationID());
+			if(!Files.exists(p2)) // Checks if directory of stationID exists
+				Files.createDirectory(p2);
+			
+			Files.createFile(toSave);
+			BufferedWriter writer = Files.newBufferedWriter(toSave);
+			writer.write(raw);
+			writer.close();
+			
+		} catch (IOException e) {
+			
+			e.printStackTrace();
+		}
+	}	
 
 }
