@@ -1,11 +1,31 @@
+/*
+ * Copyright (c) 2019 Tobias Teumert
+ * All rights reserved.
+ * 
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ * 
+ * 1. Redistributions of source code must retain the above copyright notice, this
+ *    list of conditions and the following disclaimer. 
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ * 
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
 package data.provider;
 
-import java.io.BufferedWriter;
 import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.List;
 import java.util.Properties;
 import java.util.stream.Collectors;
@@ -22,7 +42,8 @@ import data.model.WeatherDataParser;
 import data.model.WeatherDataParserValidationResult;
 
 /**
- * Diese Klasse soll alle Daten aus dem Web ziehen und zur Verfügung stellen
+ * Web Data Provider<br>
+ * Fetches Data from given URL and parses the Input into the Model
  * 
  * @author Tobias Teumert
  * 
@@ -30,18 +51,28 @@ import data.model.WeatherDataParserValidationResult;
 
 public class WebDataProvider extends ChainedDataProvider {
 	
-	//private final Multimap<StationId,Sounding> soundingsByStation  = MultimapBuilder.hashKeys().arrayListValues().build();
-		
+	/** The Parser*/		
 	private WeatherDataParser parser;
+	
+	/** The applied Caching Strategy*/
+	private RawCachingStrategy strategy;
+	
 
 	public WebDataProvider(Properties properties) {
 		this.properties = properties;
 		parser = new WeatherDataParser(this::buildStation);
+		this.strategy = (String s, LocalDateTime ldt ,StationId id) -> {}; //no - op
+		
 	}
 	
-	/**
-	 * Internet Liste der Stationen
-	 */
+	public void setRawCachingStrategy(RawCachingStrategy strat)
+	{
+		this.strategy = strat;
+	}
+	
+	
+	
+	
 	@Override
 	public List<StationId> getStationsByArea(Area area) {
 
@@ -60,13 +91,7 @@ public class WebDataProvider extends ChainedDataProvider {
 				.collect(Collectors.toList());
 	}
 	
-	/**
-	 * 
-	 * @param stationId	The Station to get the Sounding from
-	 * @param time		The specific time of the Sounding to be generated
-	 * @param area      The Area in which the Station lies
-	 * @return The RawData of the specified Sounding
-	 */
+	
 	@Override
 	public Sounding getSounding(StationId station, LocalDateTime time) {
 		
@@ -87,6 +112,7 @@ public class WebDataProvider extends ChainedDataProvider {
 			
 		} catch (IOException e) {
 			e.printStackTrace();
+			//TODO Handling of no Internet connection
 			return null;
 		}
 
@@ -94,42 +120,28 @@ public class WebDataProvider extends ChainedDataProvider {
 
 		if (result.isValid()) {
 			
-			stations.add(result.getData().getStation());
+			stations.add(result.getData().getStation());			
 			
-			String datadir = ""+properties.get("data_dir");
-			
-			Path toSave = Path.of(datadir)
-					.resolve(""+station.getArea().getAreaCode())
-					.resolve(""+station.getStationID())
-					.resolve(""+time.toLocalDate().format(DateTimeFormatter.ISO_DATE)+"T"+time.getHour()+".txt");
-			try {
-				
-				Path p1 = Path.of(datadir).resolve(station.getArea().getAreaCode());
-				if(!Files.exists(p1))
-					Files.createDirectories(p1);
-				Path p2 = p1.resolve(""+station.getStationID());
-				if(!Files.exists(p2))
-					Files.createDirectory(p2);
-				
-				Files.createFile(toSave);
-				BufferedWriter writer = Files.newBufferedWriter(toSave);
-				writer.write(doc.select("pre").first().text());
-				writer.close();
-				
-			} catch (IOException e) {
-				
-				e.printStackTrace();
-			}
-			
+			strategy.cacheRawSounding(doc.select("pre").first().text(), time, station);
 			return result.getData();
 		} else {
-			System.out.println("No Valid Result from Parser");
+			System.out.println("WDP - No Valid Result from Parser");
 			System.out.println(result.getException());
 			
 			result.getData().setDateAndTime(time);			
 			return result.getData();
 		}
 
+	}
+	/**
+	 * The caching strategy
+	 * @author Tobias
+	 *
+	 */
+	@FunctionalInterface
+	public static interface RawCachingStrategy{
+		
+		void cacheRawSounding(String raw,LocalDateTime ldt, StationId id);
 	}
 
 	@Override
