@@ -1,10 +1,12 @@
 package test;
 
+import java.time.LocalDate;
+
 import javafx.application.Application;
-import javafx.event.ActionEvent;
-import javafx.event.EventHandler;
 import javafx.scene.Scene;
-import javafx.scene.control.Button;
+import javafx.scene.chart.LineChart;
+import javafx.scene.chart.NumberAxis;
+import javafx.scene.chart.XYChart;
 import javafx.scene.layout.StackPane;
 import javafx.stage.Stage;
 
@@ -13,22 +15,29 @@ public class Test extends Application {
 		launch(args);
 	}
 
+	@SuppressWarnings({ "unchecked", "rawtypes" })
 	@Override
 	public void start(Stage primaryStage)
 	{
 		primaryStage.setTitle("Hello World!");
-		Button btn = new Button();
-		btn.setText("Say 'Hello World'");
-		btn.setOnAction(new EventHandler<ActionEvent>() {
-
-			@Override
-			public void handle(ActionEvent event) {
-				System.out.println("Hello World!");
-			}
-		});
-
 		StackPane root = new StackPane();
-		root.getChildren().add(btn);
+		
+		//LocalDateAxis xAxis = new LocalDateAxis();
+		NumberAxis xAxis = new NumberAxis();
+		NumberAxis yAxis = new NumberAxis();
+		
+		LineChart lineChart = new LineChart<Number, Number>(xAxis, yAxis);
+		
+		XYChart.Series<LocalDate, Number> series = new XYChart.Series();
+        series.setName("My portfolio");
+		
+		series.getData().add(new XYChart.Data(12, 23D));
+		series.getData().add(new XYChart.Data(15, 28D));
+		
+		lineChart.getData().add(series);
+		
+		root.getChildren().add(lineChart);
+		
 		primaryStage.setScene(new Scene(root, 300, 250));
 		primaryStage.show();
 	}
